Halt support (and chr support in BlitzMax)
Please note Halt will be a very dangerous command to use, depending on the chosen platform, as it will always end the program regardless of the current state.
In a "protected" language such a BlitzMax or Go where a strong garbage collector is present and properly executed first, probably no porblem, but if you have a module exporting to C/C++ this can and WILL lead to memory leaks if not used with care.